ai_service/requirements.txt
fastapi
uvicorn
pydantic
PyPDF2
python-docx
openai
python-dotenv
python-multipart
